fastapi_aigram_app/app/requirements-dev.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.12.14
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
dnspython==2.7.0
docopt==0.6.2
dulwich==0.22.7
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.7
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.17.0
flake8==7.1.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
installer==0.7.0
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.6.0
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.1
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pkginfo==1.12.0
platformdirs==4.3.6
poetry-core==2.0.1
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2==2.9.10
pure_eval==0.2.3
pycodestyle==2.12.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydentic==0.0.1.dev3
pyflakes==3.2.0
Pygments==2.19.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-stdnum==1.20
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.45.2
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
trove-classifiers==2025.1.15.22
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.1
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
yarg==0.1.9
yarl==1.18.3
